combine ActorBase and helpers_actor.py into actors/base.py
